Converted to Discord.JS v12, starting database

+++ commands/poweroff.js
@@ -17,26 +17,30 @@
  *   along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  * *************************************/
-module.exports.run = async (client, message, args) => {
-  const Discord = require('discord.js');
-  if (!['242775871059001344'].includes(message.author.id)) return message.reply('Nope! You need the person who created this bot to use this command.');
-  const embed = new Discord.RichEmbed()
-      .setAuthor('AleeBot Status', client.user.avatarURL)
-      .setDescription('AleeBot is now going offline...')
-      .setColor('#ff3333');
-  await client.channels.find('id', '606602551634296968').send({embed});
-  await message.reply(':warning: AleeBot will now exit!');
-  client.destroy();
-  process.exit(0);
+module.exports.run = async (client, message) => {
+	const Discord = require('discord.js');
+	if (!['242775871059001344'].includes(message.author.id)) return message.reply('Nope! You need the person who created this bot to use this command.');
+	const stopEmbed = new Discord.MessageEmbed()
+		.setAuthor('AleeBot Status', client.user.avatarURL())
+		.setDescription('AleeBot is now going offline...')
+		.setColor('#ff3333');
+	
+	let statusChannel = client.channels.cache.get('606602551634296968');
+	if (!statusChannel) return console.error('The status channel does not exist! Skipping.');
+	await statusChannel.send(stopEmbed);
+	await message.reply(':warning: AleeBot will now exit!');
+	console.log('[i] AleeBot will now exit!'.blue);
+	client.destroy();
+	process.exit(0);
 };
 
 exports.conf = {
-  aliases: ['reboot'],
-  guildOnly: false,
+	aliases: ['reboot'],
+	guildOnly: false,
 };
 exports.help = {
-  name: 'poweroff',
-  description: 'Turns off AleeBot.',
-  usage: 'poweroff',
-  category: '- Owners Only',
+	name: 'poweroff',
+	description: 'Turns off AleeBot.',
+	usage: 'poweroff',
+	category: '- Owners Only',
 };
